doc/requirements.txt
sphinx
sphinx-rtd-dark-mode
breathe
jupytext
markdown
myst_parser
pyyaml